Added present and absent list though incomplete
Added the list of students In attendance above the full roster list.
At this time it is unclear as of to why not all students show as absent.

This was added in a rather crude manner unfortunately. I will trying to clean it
later on with making a _present.html.erb in my  attendances view and rendering it
though it was rather unsuccessful when I tried that with my exising index page. As
a result I moved the code.